Change structure of list project so that it can be included in other projects
